refactor:#167 modal 명칭 변경
