+++ src/app/(client)/meal/post/page.tsx
@@ -1,4 +1,4 @@
-import Contents from '@/app/(client)/meal/post/components/contents';
+import Contents from '@/app/(client)/meal/post/components/upload-image-section';
 import { getFoodAnalysisDetail } from '@/lib/apis/analysis-request.api';
 
 const MealPostPage = async (): Promise<JSX.Element> => {
